requirements.txt
paddlepaddle==2.2.2
paddleocr==2.2
openvino-dev==2022.1.0.dev20220302